% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdtable2df.R
\name{mdtbl2df}
\alias{mdtbl2df}
\title{Convert markdown table to data.frame}
\usage{
mdtbl2df(string)
}
\arguments{
\item{string}{text}
}
\description{
Convert markdown table to data.frame
}
\examples{
md_table <- c(
"|hoge|hage|",
"|:-:|--:|",
"|1.0|a|",
"|2|abc|"
)
mdtbl2df(md_table)

\dontrun{
|hoge|hage|hige|
|:-:|--:|:-:|
|1.0|a  |1L|
|2| abc |foo|

clipr::read_clip() \%>\%
  mdtbl2df() \%>\%
  knitr::kable(align = "lrc") \%>\%
  clipr::write_clip()
}

}
