% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdtable2df.R
\name{mdtbl2mdtbl}
\alias{mdtbl2mdtbl}
\title{Convert markdown table to markdown table}
\usage{
mdtbl2mdtbl(string, ...)
}
\arguments{
\item{string}{text}

\item{...}{...}
}
\description{
Convert markdown table to markdown table
}
