% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_helper.R
\name{replace_all_celcius}
\alias{replace_all_celcius}
\title{Replace all "℃" to tex siunitx command}
\usage{
replace_all_celcius(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "℃" to tex siunitx command
}
\examples{
replace_all_celcius("20℃")

}
