% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_g.R
\name{replace_all_g}
\alias{replace_all_g}
\title{Replace all "g" to tex siunitx command}
\usage{
replace_all_g(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "g" to tex siunitx command
}
\examples{
replace_all_g("20.0 g")

}
