% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_g.R
\name{replace_all_kg}
\alias{replace_all_kg}
\title{Replace all "kg" to tex siunitx command}
\usage{
replace_all_kg(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "kg" to tex siunitx command
}
\examples{
replace_all_kg("20.0 kg")

}
