% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_L.R
\name{replace_all_L}
\alias{replace_all_L}
\title{Replace all "L" to tex siunitx command}
\usage{
replace_all_L(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "L" to tex siunitx command
}
\examples{
replace_all_L("20.0 L")

}
