% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_M.R
\name{replace_all_M}
\alias{replace_all_M}
\title{Replace all "M" to tex siunitx command}
\usage{
replace_all_M(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "M" to tex siunitx command
}
\examples{
replace_all_M("20.0 M")

}
