% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_g.R
\name{replace_all_mg}
\alias{replace_all_mg}
\title{Replace all "mg" to tex siunitx command}
\usage{
replace_all_mg(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "mg" to tex siunitx command
}
\examples{
replace_all_mg("20.0 mg")

}
