% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_L.R
\name{replace_all_mL}
\alias{replace_all_mL}
\title{Replace all "mL" to tex siunitx command}
\usage{
replace_all_mL(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "mL" to tex siunitx command
}
\examples{
replace_all_mL("20.0 mL")

}
