% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_M.R
\name{replace_all_mM}
\alias{replace_all_mM}
\title{Replace all "mM" to tex siunitx command}
\usage{
replace_all_mM(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "mM" to tex siunitx command
}
\examples{
replace_all_mM("20.0 mM")

}
