% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_g.R
\name{replace_all_ng}
\alias{replace_all_ng}
\title{Replace all "ng" to tex siunitx command}
\usage{
replace_all_ng(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "ng" to tex siunitx command
}
\examples{
replace_all_ng("20.0 ng")

}
