% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_L.R
\name{replace_all_nL}
\alias{replace_all_nL}
\title{Replace all "nL" to tex siunitx command}
\usage{
replace_all_nL(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "nL" to tex siunitx command
}
\examples{
replace_all_nL("20.0 nL")

}
