% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_M.R
\name{replace_all_nM}
\alias{replace_all_nM}
\title{Replace all "nM" to tex siunitx command}
\usage{
replace_all_nM(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "nM" to tex siunitx command
}
\examples{
replace_all_nM("20.0 nM")

}
