% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_g.R
\name{replace_all_pg}
\alias{replace_all_pg}
\title{Replace all "pg" to tex siunitx command}
\usage{
replace_all_pg(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "pg" to tex siunitx command
}
\examples{
replace_all_pg("20.0 pg")

}
