% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_L.R
\name{replace_all_pL}
\alias{replace_all_pL}
\title{Replace all "pL" to tex siunitx command}
\usage{
replace_all_pL(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "pL" to tex siunitx command
}
\examples{
replace_all_pL("20.0 pL")

}
