% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_M.R
\name{replace_all_pM}
\alias{replace_all_pM}
\title{Replace all "pM" to tex siunitx command}
\usage{
replace_all_pM(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "pM" to tex siunitx command
}
\examples{
replace_all_pM("20.0 pM")

}
