% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_helper.R
\name{replace_all_siunit}
\alias{replace_all_siunit}
\title{Replace all SI unit to tex siunitx command}
\usage{
replace_all_siunit(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all SI unit to tex siunitx command
}
\examples{
replace_all_uL("20.0 uL")

}
