% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_helper.R
\name{replace_all_ug}
\alias{replace_all_ug}
\title{Replace all "ug" to tex siunitx command}
\usage{
replace_all_ug(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "ug" to tex siunitx command
}
\examples{
replace_all_ug("20.0 ug")

}
