% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_helper.R
\name{replace_all_ugmL}
\alias{replace_all_ugmL}
\title{Replace all "ug/mL" to tex siunitx command}
\usage{
replace_all_ugmL(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "ug/mL" to tex siunitx command
}
\examples{
replace_all_ugmL("20.0 ug/mL")

}
