% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_helper.R
\name{replace_all_uL}
\alias{replace_all_uL}
\title{Replace all "uL" to tex siunitx command}
\usage{
replace_all_uL(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "uL" to tex siunitx command
}
\examples{
replace_all_uL("20.0 uL")

}
