% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx_M.R
\name{replace_all_uM}
\alias{replace_all_uM}
\title{Replace all "uM" to tex siunitx command}
\usage{
replace_all_uM(string)
}
\arguments{
\item{string}{text}
}
\description{
Replace all "uM" to tex siunitx command
}
\examples{
replace_all_uM("20.0 uM")

}
